backend/requirements.txt
requests>=2.31.0
pandas>=2.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
numpy>=1.24.0
curl_cffi>=0.5.10
dhanhq>=1.3.0
